# docs/swagger.yaml
definitions:
  handler.ShortenUrlInputBody:
    properties:
      exp:
        example: 3600000000000
        minimum: 0
        type: integer
      url:
        example: https://example.com/long-page
        type: string
    required:
    - exp
    - url
    type: object
  model.HealthCheck:
    properties:
      instance_id:
        type: string
      message:
        type: string
      redis_status:
        type: string
      service_name:
        type: string
    type: object
  model.ShortenUrlResponse:
    properties:
      code:
        example: a1b2c3d
        type: string
      message:
        example: Shorten URL generated successfully!
        type: string
    type: object
info:
  contact: {}
  description: Bookmark management service API
  title: Bookmark Management API
  version: "1.0"
paths:
  /health-check:
    get:
      consumes:
      - application/json
      description: check health of the service
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.HealthCheck'
      summary: check health of the service
      tags:
      - health-check
  /links/redirect/{code}:
    get:
      description: Redirect to long URL by code
      parameters:
      - description: Code
        in: path
        name: code
        required: true
        type: string
      responses:
        "302":
          description: Redirect to long URL
          schema:
            type: string
        "400":
          description: Invalid input
          schema:
            type: string
        "404":
          description: Code not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Redirect to long URL
      tags:
      - links
  /links/shorten:
    post:
      consumes:
      - application/json
      description: Create a short code for a long URL and store it in Redis until
        expire (duration in nanoseconds, e.g. 3600000000000 = 1 hour)
      parameters:
      - description: URL to shorten and TTL
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.ShortenUrlInputBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ShortenUrlResponse'
        "400":
          description: Invalid input
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create shorten url
      tags:
      - links
swagger: "2.0"